fix: prevent unauthorized access
Previously, it was possible to access pages without being
logged in. Now, if an unauthenticated user tries to access
a page that requires authentication, he will be redirected
to the login page and then redirected back to where he was
trying to access.

The same will happen if ones credentials are in 'localStorage'.
Except that the login page will prevent input and will load
the previously saved credentials.